refactor: -Wall -Wextra -Werror flags added

+++ mandatory/srcs/draw_sky_n_ground.c
@@ -6,7 +6,7 @@
 /*   By: Andie <Andie@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/08 19:40:18 by Andie             #+#    #+#             */
-/*   Updated: 2025/05/17 13:05:31 by acaceres         ###   ########.fr       */
+/*   Updated: 2025/05/22 19:48:23 by acaceres         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ void	draw_sky_and_ground(t_cub *cub, int wall_top,
 		i++;
 	}
 	i = wall_bottom;
-	while (i < cub->main_window->res.height)
+	while ((unsigned int)i < cub->main_window->res.height)
 	{
 		pixel.py = (float)i;
 		pixel.color = cub->floor;
